SMITE Calibration

Create a model (x) based on the 'forward' coral variable matrix (A) and the reconstruction target (b).

Usage

SMITE.calib(
  A,
  b,
  Ae = NULL,
  be = NULL,
  it = 10000,
  noise = "white",
  acc = NULL,
  eigenclean = NULL,
  alpha = 0.05,
  xval = NULL
)

Arguments

A

The 'forward' coral variable matrix (t x p; time in rows, coral parameters in columns).

b

The reconstruction target (t x 1).

Ae

A t x p matrix containing the errors for the forward matrix.

be

Error estimates for the reconstruction target. It can either be one value or a vector of equal length to the reconstruction target (t x 1).

it

The number of bootstrap Monte Carlo iterations (recommended at 10,000).

noise

Either 'white' or 'red', describing how the noise is propagated into the model.

acc

Autocorrelation Coefficient; if noise is specified as 'red', it describes the degree of autocorrelation in the error term.

eigenclean

Describes how (if at all) the singular values should be truncated to 'clean' the inverse solution. If the value is between 0 and 1, it will remove singular values based on the cumulative variance explained. If the value is greater than 1, it will return that many singular values (from the highest order).

alpha

The significance level for the confidence interval (i.e., 0.05 = 95-percent confidence).

Value

S - Confidence intervals on the singular values of the forward matrix,

x - SMITE model parameters, or loadings, for each column of the forward matrix.

bhat - Predicted target anomalies

recon - Predicted absolute target values

e - Relevant error metrics. It enumerates (half) the distance of the 95-percent confidence interval (the average distance from the mean to the upper and lower bounds) for the singular values and the model parameters. It also enumerates the Standard Error of Prediction (SEP), Root-Mean-Square-Error (RMSE), and the correlation coefficient (r).

Examples

# Load data from Hughes et al. (in revision)
library(SMITER)
data(BMDA_1B_Comp)
data(BMDA_1B_EComp)

# Designate forward matrix (A), reconstruction target (b), and corresponding errors (Ae, be).

proxy_names <- names(bmda_1b_comp[,c(5:ncol(bmda_1b_comp))])

A <- bmda_1b_comp[,proxy_names]
Ae <- bmda_1b_ecomp[,proxy_names]
b <- bmda_1b_comp[,'Temp']
be <- rep(0.02, length(b))

# Execute SMITE Calibration with no truncation (i.e., all information retained).
result <- SMITE.calib(A = A, b = b, Ae = Ae, be = be, eigenclean = ncol(A))
#> [1] "No window size for cross-validation (xval) specified. Using minimum window size of 1."