SMITE Reconstruction

Apply a SMITE model (m) to a 'forward' coral variable matrix (A) to produce predicted values for a reconstruction target (bhat).

SMITE.recon(
  A,
  m,
  Amu,
  Asd,
  bmu,
  bsd
)

Arguments

A

The 'forward' coral variable matrix (t x p; time in rows, coral parameters in columns).

m

A vector of SMITE model parameters (length p), ideally created from the SMITE.calib() function.

Amu

A vector of length p giving the mean values of each coral variable in the calibration dataset.

Asd

A vector of length p giving the standard deviations of each coral variable in the calibration dataset.

bmu

A single-element atomic vector giving the mean of the reconstruction target in the calibration dataset.

bsd

A single-element atomic vector giving the standard deviation of the reconstruction target in the calibration dataset.

Value

Returns a vector of predicted reconstruction target values (bhat). Minimum errors (i.e., not included age model errors) for bhat can be estimated from the SEP calculated in the SMITE.calib() function.

References

Hughes, H. P., Thompson, D. M., Foster, G. L., Lees, J., Surge, D. (in revision). “Synthetic and Practical Reconstructions of SST and pH Using the Novel Multiproxy SMITE Method.” PLOS ONE.

Author

Hunter P. Hughes

Examples

# Load data from Hughes et al. (in revision)
library(SMITER)
data(BMDA_1B_Comp)
data(BMDA_1B_EComp)

# Designate forward matrix (A), reconstruction target (b), and corresponding errors (Ae, be).

proxy_names <- names(bmda_1b_comp[,c(5:ncol(bmda_1b_comp))])

A <- bmda_1b_comp[,proxy_names]
Ae <- bmda_1b_ecomp[,proxy_names]
b <- bmda_1b_comp[,'Temp']
be <- rep(0.02, length(b))

# Execute SMITE Calibration with no truncation (i.e., all information retained).

result <- SMITE.calib(A = A, b = b, Ae = Ae, be = be, eigenclean = ncol(A))

# Extract SMITE model parameters #
m <- result$m$m.mu

# Reconstruction (should be identical to result$recon)
bhat <- SMITE.recon(
  A = A,
  m = m,
  Amu = colMeans(A),
  Asd = apply(A, 2, function(x) sd(x)),
  bmu = mean(b),
  bsd = sd(b)
)

result$recon
#>    bhat.low  bhat.mu bhat.high
#> 1  27.05325 28.86139  30.55547
#> 2  24.71045 26.49063  28.27164
#> 3  24.92474 26.53376  28.14986
#> 4  23.86895 25.40753  26.95545
#> 5  22.06896 23.54265  24.98656
#> 6  19.70709 21.46609  23.15375
#> 7  19.07819 20.94818  22.80992
#> 8  18.89890 20.51567  22.15208
#> 9  21.06553 22.59325  24.06295
#> 10 23.17642 24.86350  26.47177
#> 11 24.97058 26.85822  28.71285
#> 12 26.27302 27.97278  29.63151
#> 13 26.81256 28.35513  29.84164
#> 14 26.00033 27.59611  29.19236
#> 15 24.81957 26.38086  27.85224
#> 16 23.37819 24.90373  26.43708
#> 17 20.66512 22.23570  23.85933
#> 18 18.71349 20.33556  22.02138
#> 19 18.95506 20.44850  22.03285
#> 20 19.50010 21.22034  22.97347
#> 21 20.86041 22.57487  24.25593
#> 22 20.23048 21.89511  23.60120
#> 23 22.32870 23.90824  25.50206
#> 24 25.72807 27.44491  29.12771
#> 25 26.70167 28.26452  29.83762
#> 26 26.42080 27.74512  29.11135
#> 27 24.26558 25.78140  27.30915
#> 28 21.93088 23.35733  24.83474
#> 29 20.49602 21.96019  23.47767
#> 30 18.81556 20.16831  21.54888
#> 31 18.38739 19.79337  21.24534
#> 32 18.63018 20.16216  21.74570
#> 33 19.61190 21.08338  22.58242
bhat
#>             [,1]
#>  [1,]  8.4763991
#>  [2,]  5.6120389
#>  [3,]  6.1109454
#>  [4,]  4.9887220
#>  [5,]  3.1031210
#>  [6,]  0.9310856
#>  [7,]  0.2564495
#>  [8,] -0.3141264
#>  [9,]  2.0222248
#> [10,]  4.6837076
#> [11,]  6.5063441
#> [12,]  7.6847079
#> [13,]  8.1875984
#> [14,]  7.0239930
#> [15,]  6.1909385
#> [16,]  4.1560462
#> [17,]  1.3135035
#> [18,] -0.8350191
#> [19,] -0.7092167
#> [20,]  0.3239517
#> [21,]  2.1122135
#> [22,]  1.1407826
#> [23,]  3.3840137
#> [24,]  7.2162294
#> [25,]  7.8909197
#> [26,]  7.5013972
#> [27,]  5.3224716
#> [28,]  2.5969831
#> [29,]  1.0779108
#> [30,] -0.8841770
#> [31,] -1.1664231
#> [32,] -0.5116102
#> [33,]  0.3487934